feat: add sentinel supervisor with binary watching and auto-restart
Refactor taskguild-agent into subcommands using kingpin CLI framework:
- "run" subcommand contains the original agent manager logic
- "sentinel" subcommand supervises the run process with automatic
  restart on crash (with exponential backoff) and binary file change
  detection via fsnotify for zero-downtime deployments
